+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [1.7.0] - 2025-08-25
 
+### Added
+- Workflow generator
+
 ### Changes
 - Port over `cmdx-parallel` changes
 - Port over `cmdx-i18n` changes